==> Arquivos/07. Google BigQuery/8.executar comando antes da carga.ipynb <==
import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq
from datetime import datetime

df = pd.read_excel(r"C:\Users\edmil\OneDrive\Documentos\Material Python-Google BigQuery\Arquivos\Origem\arquivos_excel\Ordens.xlsx")
df.head()

tipos_de_dados = df.dtypes
print(tipos_de_dados)


# filto no meu dataflame antes de realizar a carga (apenas o mês e ano atuais)
"""
hoje = datetime.now()
mes_atual = hoje.month
ano_atual = hoje.year

df_filtrado = df[(df['created_at'].dt.month == mes_atual) & (df['created_at'].dt.year == ano_atual)]


"""

#hoje = datetime.now().year
#print(hoje)

#df_filtrado.head(5)


hoje = datetime.now()
mes_atual = hoje.month
ano_atual = hoje.year

df = df[(df['created_at'].dt.month == mes_atual) & (df['created_at'].dt.year == ano_atual)]

df.head()



# Especifique o caminho para o arquivo JSON de credenciais
credencial = r"C:\Users\edmil\Downloads\pythoncurso-398511-260383f13f39.json"

# Crie um cliente BigQuery usando as credenciais
clientacao = bigquery.Client.from_service_account_json(credencial)

comandoapagar = """ 

 delete
 FROM `pythoncurso-398511.AulaPython.Ordens`
 where FORMAT_TIMESTAMP('%Y', created_at)  = FORMAT_TIMESTAMP('%Y', CURRENT_DATE())
 and  FORMAT_TIMESTAMP('%m', created_at)  = FORMAT_TIMESTAMP('%m', CURRENT_DATE())


"""
# Execute a consulta
apagardados = clientacao.query(comandoapagar)

https://cloud.google.com/bigquery/docs/reference/standard-sql/format-elements


# credenciais para validar importar dados para o google bigquery
CredImportacao = service_account.Credentials.from_service_account_file(
    credencial,
    scopes=['https://www.googleapis.com/auth/bigquery']
)

#variaveis do projeto 
projeto = 'pythoncurso-398511'
dataset = 'AulaPython'
tabela = 'Ordens'
parameretro ='append'      
#replace = subistuir a antiga pela original 
#append = adiciona dados no fim da tabela 

df.to_gbq(destination_table=f'{projeto}.{dataset}.{tabela}',   
          project_id=projeto,
          if_exists=parameretro,
          credentials=CredImportacao)

"""  

SELECT *
,extract(YEAR  FROM created_at) as ano
,extract(month  FROM created_at) as mes
,FORMAT_TIMESTAMP('%Y', created_at) AS anoFormat
,FORMAT_TIMESTAMP('%m', created_at) AS mesformat
,CURRENT_DATE() hoje
 FROM `pythoncurso-398511.AulaPython.Ordens` 


"""

#Delete
"""

delete FROM `pythoncurso-398511.AulaPython.Ordens` 
 where FORMAT_TIMESTAMP('%Y', created_at) = FORMAT_TIMESTAMP('%Y', CURRENT_DATE())
 and  FORMAT_TIMESTAMP('%m', created_at) = FORMAT_TIMESTAMP('%m', CURRENT_DATE())
"""
